+++ scripts/RT/harvest_wb.py
@@ -87,14 +87,14 @@ def load_precip():
       dict of [date][huc12]
     """
     # 1. Build GeoPandas DataFrame of HUC12s of interest
-    idep = get_dbconn("idep")
+    idep = get_dbconn("dep")
     huc12df = gpd.GeoDataFrame.from_postgis(
         """
-        SELECT huc_12, ST_Transform(simple_geom, 4326) as geo
-        from huc12 WHERE scenario = 0
+        SELECT huc12_code, ST_Transform(simple_geom, 4326) as geo
+        from huc12 WHERE scenario_id = 0
     """,
         idep,
-        index_col="huc_12",
+        index_col="huc12_code",
         geom_col="geo",
     )
     # 2. Loop over dates